adding value to CONN-file
if you save a project, the CONN-file saves the current value (position) of the instruments.
if you open the project the saved value is set for the instruments.
Problem: if you use the same instrument page in many projects, you have different values and so you need different connection files