handling read/write with syncdirection properly
reading which parameters are requested. Only parameters which are connected to instruments are requested. In return microcontroller sends the currently stored value to python.
python takes the value und sets the value of the instruments.
You are able to change the instrument positions and press syncdirection to change to write mode.
if done so, the set values are written instantly onto the microcotroller