external parametersets + correction of saving messages + only values in saved single set params + New example projected
parameter values are handled external. when saving a project , two files are generated in the same location as the project itself.
if i am in a project and press save then the connected paramfiles are updated.
if i am not in a project (untitled) and press save, a single set param file is generated.
If I press save params as: single set param file is generated  <-- no matter in which case.
new example project contains proper description of what parameters of arduino are connected.