haney typechanges and different types of instrWidgets + converting everything to coherent datatype of param + min and max are dependent on value + param is master for min and max
Parameters are master of Min and Max. Min and Max are now editable again. The user can set Min and Max. If a Parameter is connected to a instrWidget. The instrWidget takes the Min and Max of Param. If a uint8 or int16 param is connected to double spinbox the stepvalue changes to 1. if the doublespinbox is connected back to a float, the initial set stepvalue is set again.
If a type of a connected param is changed, param gets disconnected from instrWidget.
if i choose param dataType float --> uint8 or int16 , min max and values are converted to int.